backend/requirement.txt
fastapi==0.104.1
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
pwdlib==0.2.0
python-multipart==0.0.6
alembic==1.13.0
psycopg2-binary==2.9.9
uvicorn==0.24.0
cloudinary==1.39.1
pillow==10.3.0
pytesseract==0.3.10
